catastrophe/pyproject.toml
[project]
name = "catastrophe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "datasets>=4.4.2",
    "google-genai>=1.57.0",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
]
